Implement VMM public interface on riscv64
