+++ kernel/arch/riscv64/vmm/vmmapper.c
@@ -4,28 +4,8 @@
  *
  * Copyright (c) 2025 Ross Bamford
  * 
- * First time in this file? Well, buckle up, buttercup - you're in
- * for a bit of a ride...
- * 
- * On RISC-V, we use a direct mapping, since recursive isn't possible
- * on that platform. This is used for everything, including management
- * of page tables themselves, which presents us a bit of a 
- * chicken-and-egg situation when it comes to actually building the 
- * direct map during init.
- * 
- * A lot of the complexity in here is dealing with that - once the 
- * direct map is built, it makes the rest of VMM a breeze compared
- * to x86_64.
- * 
- * I try to be as efficient as possible with the mapping, using the 
- * largest tables possible for each region and breaking it into 
- * naturally-aligned blocks for larger sizes where possible.
- * 
- * Overhead will vary, on qemu with 8GiB RAM, it currently needs 
- * just over 32KiB for the direct-mapping tables (mapping all of 
- * physical RAM).
- * 
- * There's more info in MemoryMap.md.
+ * Easy-peasy, based on the direct mapping. The hard work to set
+ * that up happens in vmmapper_init.c :)
  */
 
 #include <stddef.h>
@@ -35,8 +15,10 @@
 #include "pmm/pagealloc.h"
 #include "riscv64/kdrivers/cpu.h"
 #include "riscv64/vmm/vmconfig.h"
-#include "riscv64/vmm/vmmapper.h"
+#include "spinlock.h"
 #include "std/string.h"
+
+#include "vmm/vmmapper.h"
 
 #ifdef DEBUG_VMM
 #include "kprintf.h"
@@ -51,1061 +33,211 @@
 #define vdebugf(...)
 #endif
 
-#define MAX_PHYS_ADDR (((size_t)127 * 1024 * 1024 * 1024 * 1024)) // 127 TiB
+#ifdef UNIT_TESTS
+#define STATIC_EXCEPT_TESTS
+#else
+#define STATIC_EXCEPT_TESTS static
+#endif
 
-// Physical memory region
+typedef enum {
+    PT_LEVEL_PML4 = 4,
+    PT_LEVEL_PDPT = 3,
+    PT_LEVEL_PD = 2,
+    PT_LEVEL_PT = 1,
+} PagetableLevel;
+
 extern MemoryRegion *physical_region;
 
-static inline uintptr_t per_cpu_temp_base_vaddr_for_pt_index(uint16_t index) {
-    return PER_CPU_TEMP_PAGE_BASE + (index * PAGE_SIZE);
+static SpinLock vmm_map_lock;
+
+// TODO locking in here is very coarse-grained - it could be done based
+//      on the top-level table instead, for example...
+//
+
+/*
+ * Returns true if the given entry is a leaf, i.e.
+ * has any of the READ, WRITE or EXEC bits set.
+ */
+STATIC_EXCEPT_TESTS inline bool is_leaf(uint64_t table_entry) {
+    return table_entry & (PG_READ | PG_WRITE | PG_EXEC);
 }
 
-static inline void *per_cpu_temp_base_vptr_for_pt_index(uint16_t index) {
-    return (void *)per_cpu_temp_base_vaddr_for_pt_index(index);
-}
-
-static inline void map_readwrite_and_flush_offs(uint64_t *table,
-                                                uint16_t table_index,
-                                                uintptr_t base_paddr,
-                                                uintptr_t base_vaddr,
-                                                uintptr_t base_offs) {
-
-    vdebugf("map_readwrite_and_flush: table @ 0x%016lx[%d]: vaddr: "
-            "0x%016lx+0x%016lx => paddr: 0x%016lx\n",
-            (uintptr_t)table, table_index, base_vaddr, base_offs, base_paddr);
-
-    table[table_index] = vmm_phys_and_flags_to_table_entry(
-            base_paddr, PG_READ | PG_WRITE | PG_PRESENT);
-    cpu_invalidate_tlb_addr(base_vaddr + base_offs);
-    vdebugf("map_readwrite_and_flush: done, and done\n");
-}
-
-static inline void map_readwrite_and_flush(uint64_t *table,
-                                           uint16_t table_index,
-                                           uintptr_t paddr, uintptr_t vaddr) {
-    map_readwrite_and_flush_offs(table, table_index, paddr, vaddr, 0);
-}
-
-static inline void unmap_and_flush(uint64_t *table, uint16_t table_index,
-                                   uintptr_t vaddr) {
-    table[table_index] = 0;
-    cpu_invalidate_tlb_addr(vaddr);
-}
-
-static void vmm_init_map_terapage(uint64_t *pml4, uintptr_t base,
-                                  uint8_t flags) {
-    vdebugf("vmm_init_map_terapage: Mapping phys 0x%016lx with length %ld into "
-            "PML4 @ 0x%016lx\n",
-            base, TERA_PAGE_SIZE, (uintptr_t)pml4);
-
-    if ((base + TERA_PAGE_SIZE) > MAX_PHYS_ADDR) {
-        debugf("WARN: Refusing to map memory at 0x%016lx [%ld bytes] due to "
-               "address-space overflow\n",
-               base, TERA_PAGE_SIZE);
-        return;
+/*
+ * Ensure tables are mapped to the specified level.
+ *
+ * levels (1-based to fit with PML4 naming):
+ *
+ *   4 - Ensures PML4
+ *   3 - Ensures PDPT
+ *   2 - Ensures PD
+ *   1 - Ensures PT
+ * 
+ * Returns a virtual pointer to the table of the specified 
+ * level, which may be newly created.
+ * 
+ * Must be called with VMM locked!
+ */
+STATIC_EXCEPT_TESTS uint64_t *ensure_tables(uint64_t *root_table,
+                                            uintptr_t virt_addr,
+                                            PagetableLevel to_level) {
+    if (to_level < 1 || to_level > 4) {
+        return NULL;
     }
 
-    uintptr_t vaddr = vmm_phys_to_virt(base);
-    vdebugf("  -> vaddr is 0x%016lx\n", vaddr);
+    int levels_remain = 4 - to_level;
+    uint64_t *current_table = root_table;
 
-    uint16_t pml4_index = vmm_virt_to_pml4_index(vaddr);
+    while (levels_remain) {
+        uint16_t current_index =
+                vmm_virt_to_table_index(virt_addr, levels_remain + 1);
+        uint64_t current_entry = current_table[current_index];
+
+        if ((current_entry & PG_PRESENT) == 0) {
+            uintptr_t new_table = page_alloc(physical_region);
+
+            if (new_table & 0xfff) {
+                // TODO unmap what we mapped so far? Is it worth it?
+                return NULL;
+            }
+
+            uint64_t *new_ptr = vmm_phys_to_virt_ptr(new_table);
+
+            memclr(new_ptr, VM_PAGE_SIZE);
+
+            current_table[current_index] =
+                    vmm_phys_and_flags_to_table_entry(new_table, PG_PRESENT);
+            current_table = new_ptr;
+        } else {
+            // special case - if this is a leaf, but we wanted a lower level,
+            // we'll just fail...
+            if (levels_remain > 1 && is_leaf(current_entry)) {
+                return NULL;
+            }
+
+            current_table = vmm_phys_to_virt_ptr(
+                    vmm_table_entry_to_phys(current_entry));
+        }
+
+        levels_remain -= 1;
+    }
+
+    return current_table;
+}
+
+inline void vmm_invalidate_page(uintptr_t virt_addr) {
+    cpu_invalidate_tlb_addr(virt_addr);
+}
+
+inline bool vmm_map_page_containing_in(uint64_t *pml4, uintptr_t virt_addr,
+                                       uint64_t phys_addr, uint16_t flags) {
+    uint64_t lock_flags = spinlock_lock_irqsave(&vmm_map_lock);
+
+    virt_addr &= PAGE_ALIGN_MASK;
+
+    uint64_t *pt = ensure_tables(pml4, virt_addr, PT_LEVEL_PT);
+
+    if (!pt) {
+        spinlock_unlock_irqrestore(&vmm_map_lock, lock_flags);
+        return false;
+    }
+
+    pt[vmm_virt_to_pt_index(virt_addr)] =
+            vmm_phys_and_flags_to_table_entry(phys_addr, flags | PG_PRESENT);
+
+    vmm_invalidate_page(virt_addr);
+
+    spinlock_unlock_irqrestore(&vmm_map_lock, lock_flags);
+    return true;
+}
+
+inline bool vmm_map_page_containing(uintptr_t virt_addr, uint64_t phys_addr,
+                                    uint16_t flags) {
+    return vmm_map_page_containing_in(
+            vmm_phys_to_virt_ptr(cpu_satp_to_root_table_phys(cpu_read_satp())),
+            virt_addr, phys_addr, flags);
+}
+
+inline bool vmm_map_page_in(uint64_t *pml4, uintptr_t virt_addr, uint64_t page,
+                            uint16_t flags) {
+    return vmm_map_page_containing_in(pml4, virt_addr, page, flags);
+}
+
+inline bool vmm_map_page(uintptr_t virt_addr, uint64_t page, uint16_t flags) {
+    return vmm_map_page_containing(virt_addr, page, flags);
+}
+
+uintptr_t vmm_unmap_page_in(uint64_t *pml4, uintptr_t virt_addr) {
+    uint64_t lock_flags = spinlock_lock_irqsave(&vmm_map_lock);
+
+    uint16_t pml4_index = vmm_virt_to_pml4_index(virt_addr);
     uint64_t pml4e = pml4[pml4_index];
 
     if (pml4e & PG_PRESENT) {
-        debugf("WARN [BUG]: PML4 for terapage at 0x%016lx is already mapped\n",
-               vaddr);
-    }
+        if (is_leaf(pml4e)) {
+            // unmapping a terapage
+            pml4[pml4_index] = 0;
 
-    pml4[pml4e] = vmm_phys_and_flags_to_table_entry(base, flags);
-}
+            vmm_invalidate_page(virt_addr);
 
-static uint64_t *const ensure_direct_table_map(uint64_t *table,
-                                               uint16_t table_index,
-                                               uint64_t *temp_mapping_pt,
-                                               uint16_t temp_mapping_pt_index) {
-    vdebugf("ensure_direct_table_map: table 0x%016lx[%d] into temp mapping "
-            "0x%016lx[%d]\n",
-            (uintptr_t)table, table_index, (uintptr_t)temp_mapping_pt,
-            temp_mapping_pt_index);
-    const uint64_t table_entry = table[table_index];
-
-    const uintptr_t new_table_vaddr =
-            per_cpu_temp_base_vaddr_for_pt_index(temp_mapping_pt_index);
-    uint64_t *const new_table =
-            per_cpu_temp_base_vptr_for_pt_index(temp_mapping_pt_index);
-
-    if ((table_entry & PG_PRESENT) == 0) {
-        vdebugf("  -> adding a new child table at at index %d\n", table_index);
-
-        const uintptr_t new_table_paddr = page_alloc(physical_region);
-        if (new_table_paddr & 0xfff) {
-            panic("Ran out of memory while building new table for direct "
-                  "mapping");
+            spinlock_unlock_irqrestore(&vmm_map_lock, lock_flags);
+            return vmm_table_entry_to_phys(pml4e);
         }
 
-        map_readwrite_and_flush(temp_mapping_pt, temp_mapping_pt_index,
-                                new_table_paddr, new_table_vaddr);
+        uint64_t *pdpt = vmm_phys_to_virt_ptr(vmm_table_entry_to_phys(pml4e));
+        uint16_t pdpt_index = vmm_virt_to_pdpt_index(virt_addr);
+        uint64_t pdpte = pdpt[pdpt_index];
 
-        vdebugf("  -> New table is mapped; Clearing...");
+        if (pdpte & PG_PRESENT) {
+            if (is_leaf(pdpte)) {
+                // unmapping a gigapage
+                pdpt[pdpt_index] = 0;
 
-        for (int i = 0; i < PAGE_TABLE_ENTRIES; i++) {
-            new_table[i] = 0;
-        }
+                vmm_invalidate_page(virt_addr);
 
-        table[table_index] =
-                vmm_phys_and_flags_to_table_entry(new_table_paddr, PG_PRESENT);
-    } else {
-        vdebugf("  -> table already present, mapping...\n");
-        map_readwrite_and_flush(temp_mapping_pt, temp_mapping_pt_index,
-                                vmm_table_entry_to_phys(table_entry),
-                                new_table_vaddr);
-    }
-
-    return new_table;
-}
-
-static void vmm_init_map_gigapage(uint64_t *pml4, uint64_t *temp_mapping_pt,
-                                  uintptr_t base, uint8_t flags) {
-    vdebugf("vmm_init_map_gigapage: Mapping phys 0x%016lx with length %ld into "
-            "PML4 @ 0x%016lx\n",
-            base, GIGA_PAGE_SIZE, (uintptr_t)pml4);
-
-    if ((base + GIGA_PAGE_SIZE) > MAX_PHYS_ADDR) {
-        debugf("WARN: Refusing to map memory at 0x%016lx [%ld bytes] due to "
-               "address-space overflow\n",
-               base, GIGA_PAGE_SIZE);
-        return;
-    }
-
-    const uintptr_t vaddr = vmm_phys_to_virt(base);
-    vdebugf("  -> vaddr is 0x%016lx\n", vaddr);
-
-    // ## PDPT
-    const uint16_t pml4_index = vmm_virt_to_pml4_index(vaddr);
-    uint64_t *const pdpt =
-            ensure_direct_table_map(pml4, pml4_index, temp_mapping_pt, 0);
-
-    const uint16_t pdpt_index = vmm_virt_to_pdpt_index(vaddr);
-
-    vdebugf("  -> Mapping region at PDPT index %d\n", pdpt_index);
-
-    if (pdpt[pdpt_index] & PG_PRESENT) {
-        panic("Physical memory already direct mapped; region overlap or bug; "
-              "cannot continue");
-    }
-
-    pdpt[pdpt_index] = vmm_phys_and_flags_to_table_entry(base, flags);
-    unmap_and_flush(temp_mapping_pt, 0, (uintptr_t)pdpt);
-
-    vdebugf("vmm_init_map_gigapage: Region mapped successfully\n");
-}
-
-static void vmm_init_map_megapage(uint64_t *pml4, uint64_t *temp_mapping_pt,
-                                  uintptr_t base, uint8_t flags) {
-    vdebugf("vmm_init_map_megapage: Mapping phys 0x%016lx with length %ld into "
-            "PML4 @ 0x%016lx\n",
-            base, MEGA_PAGE_SIZE, (uintptr_t)pml4);
-
-    if ((base + MEGA_PAGE_SIZE) > MAX_PHYS_ADDR) {
-        debugf("WARN: Refusing to map memory at 0x%016lx [%ld bytes] due to "
-               "address-space overflow\n",
-               base, MEGA_PAGE_SIZE);
-        return;
-    }
-
-    const uintptr_t vaddr = vmm_phys_to_virt(base);
-    vdebugf("  -> vaddr is 0x%016lx\n", vaddr);
-
-    // ## PDPT
-    const uint16_t pml4_index = vmm_virt_to_pml4_index(vaddr);
-    uint64_t *const pdpt =
-            ensure_direct_table_map(pml4, pml4_index, temp_mapping_pt, 0);
-
-    // ## PD
-    const uint16_t pdpt_index = vmm_virt_to_pdpt_index(vaddr);
-    uint64_t *const pd =
-            ensure_direct_table_map(pdpt, pdpt_index, temp_mapping_pt, 1);
-
-    uint16_t pd_index = vmm_virt_to_pd_index(vaddr);
-
-    vdebugf("  -> Mapping region at PD index %d\n", pd_index);
-
-    if (pd[pd_index] & PG_PRESENT) {
-        panic("Physical memory already direct mapped; region overlap or bug; "
-              "cannot continue");
-    }
-
-    pd[pd_index] = vmm_phys_and_flags_to_table_entry(base, flags);
-    unmap_and_flush(temp_mapping_pt, 0, (uintptr_t)pdpt);
-    unmap_and_flush(temp_mapping_pt, 1, (uintptr_t)pd);
-
-    vdebugf("vmm_init_map_megapage: Region mapped successfully\n");
-}
-
-static void vmm_init_map_page(uint64_t *pml4, uint64_t *temp_mapping_pt,
-                              uintptr_t base, uint8_t flags) {
-    vdebugf("vmm_init_map_page: Mapping phys 0x%016lx with length %ld into "
-            "PML4 @ 0x%016lx\n",
-            base, PAGE_SIZE, (uintptr_t)pml4);
-
-    if ((base + PAGE_SIZE) > MAX_PHYS_ADDR) {
-        debugf("WARN: Refusing to map memory at 0x%016lx [%ld bytes] due to "
-               "address-space overflow\n",
-               base, PAGE_SIZE);
-        return;
-    }
-
-    const uintptr_t vaddr = vmm_phys_to_virt(base);
-    vdebugf("  -> vaddr is 0x%016lx\n", vaddr);
-
-    // ## PDPT
-    const uint16_t pml4_index = vmm_virt_to_pml4_index(vaddr);
-    uint64_t *const pdpt =
-            ensure_direct_table_map(pml4, pml4_index, temp_mapping_pt, 0);
-
-    // ## PD
-    const uint16_t pdpt_index = vmm_virt_to_pdpt_index(vaddr);
-    uint64_t *const pd =
-            ensure_direct_table_map(pdpt, pdpt_index, temp_mapping_pt, 1);
-
-    // ## PT
-    const uint16_t pd_index = vmm_virt_to_pd_index(vaddr);
-    uint64_t *const pt =
-            ensure_direct_table_map(pd, pd_index, temp_mapping_pt, 2);
-
-    uint16_t pt_index = vmm_virt_to_pt_index(vaddr);
-
-    vdebugf("  -> Mapping region at PT index %d\n", pt_index);
-
-    if (pt[pt_index] & PG_PRESENT) {
-        panic("Physical memory already direct mapped; region overlap or bug; "
-              "cannot continue");
-    }
-
-    pt[pt_index] = vmm_phys_and_flags_to_table_entry(base, flags);
-
-    unmap_and_flush(temp_mapping_pt, 0, (uintptr_t)pdpt);
-    unmap_and_flush(temp_mapping_pt, 1, (uintptr_t)pd);
-    unmap_and_flush(temp_mapping_pt, 3, (uintptr_t)pt);
-
-    vdebugf("vmm_init_map_page: Region mapped successfully\n");
-}
-
-static void vmm_init_map_region(uint64_t *pml4, uint64_t *temp_mapping_pt,
-                                Limine_MemMapEntry *entry, bool writeable) {
-    vdebugf("vmm_init_map_region: Mapping phys 0x%016lx with length %ld into "
-            "PML4 @ 0x%016lx\n",
-            entry->base, entry->length, (uintptr_t)pml4);
-    uint8_t flags =
-            PG_PRESENT | PG_GLOBAL | PG_READ | (writeable ? PG_WRITE : 0);
-    uintptr_t base = entry->base;
-    uintptr_t length = entry->length;
-
-    while (length > 0) {
-        if (length >= TERA_PAGE_SIZE && base % TERA_PAGE_SIZE == 0) {
-            // map one terapage
-            vmm_init_map_terapage(pml4, base, flags);
-            base += TERA_PAGE_SIZE;
-            length -= TERA_PAGE_SIZE;
-        } else if (length >= GIGA_PAGE_SIZE && base % GIGA_PAGE_SIZE == 0) {
-            // map one gigapage
-            vmm_init_map_gigapage(pml4, temp_mapping_pt, base, flags);
-            base += GIGA_PAGE_SIZE;
-            length -= GIGA_PAGE_SIZE;
-        } else if (length >= MEGA_PAGE_SIZE && base % MEGA_PAGE_SIZE == 0) {
-            // map one megapage
-            vmm_init_map_megapage(pml4, temp_mapping_pt, base, flags);
-            base += MEGA_PAGE_SIZE;
-            length -= MEGA_PAGE_SIZE;
-        } else if (length >= PAGE_SIZE && base % PAGE_SIZE == 0) {
-            // map one page
-            vmm_init_map_page(pml4, temp_mapping_pt, base, flags);
-            base += PAGE_SIZE;
-            length -= PAGE_SIZE;
-        } else {
-            debugf("vmm_init_map_region: WARN: %ld byte area < PAGE_SIZE "
-                   "wasted at 0x%016lx\n",
-                   length, base);
-            return;
-        }
-    }
-
-    vdebugf("vmm_init_map_region: phys 0x%016lx with length %ld mapped into "
-            "pml4 @ 0x%016lx\n",
-            entry->base, entry->length, (uintptr_t)pml4);
-}
-
-#define TERAPAGE_ALIGN_MASK (~(TERA_PAGE_SIZE - 1))
-
-static inline uintptr_t round_down_to_terapage(uintptr_t addr) {
-    return addr & TERAPAGE_ALIGN_MASK;
-}
-
-static inline uintptr_t offset_in_terapage(uintptr_t addr) {
-    return addr & (TERA_PAGE_SIZE - 1);
-}
-
-// this next bit is already hard enough to follow, so don't let clang-format
-// mess up the formatting...
-//
-// clang-format off
-
-// This just ensures a full set of page-tables exist for for the given temp_map_addr.
-//
-// We use this for temporarily mapping new page tables we need to create during the
-// build of the direct mapping, because of the aforementioned chicken-and-egg situation - 
-// without the direct map, the usual vmm_ functions for managing page tables don't
-// work, but we do need to be able to map pages in order to build the direct map...
-//
-// Although this is expected to be called in early boot it doesn't make
-// assumptions about the existing table layout, other than that the PML4 it's 
-// given is valid. This means that as the design progresses and inevitably changes
-// I'm (hopefully) less likely to waste time debugging weird issues before 
-// remembering that this thing exists :D
-//
-// It's a hell of a function, but because the process itself is a bit mind-bending
-// I wrote it with a focus on understandability of the actual logic we follow 
-// rather than optimising for performance or perceived 'cleanliness'.
-//
-// Worth noting also that it's only to be used before userspace is up, since it
-// (ab)uses some of the userspace PML4 mappings to do its work.
-//
-static uint64_t* ensure_temp_page_tables(uint64_t * const pml4, const uintptr_t temp_map_addr) {
-    vdebugf("ensure_temp_page_tables(0x%016lx, 0x%016lx)\n", (uintptr_t)pml4, temp_map_addr);
-
-    uint16_t pml4e = vmm_virt_to_pml4_index(temp_map_addr);
-
-    vdebugf("  Will try to get entry %d from the pml4e\n", pml4e);
-
-    // Do we have a PDPT covering the region?
-    if (unlikely((pml4[pml4e] & PG_PRESENT) == 0)) {
-        panic("Kernel space root mapping does not exist");
-    } else {
-        vdebugf("    PDPT is present, checking PDPT\n");
-        // Yes - map it as a terapage in low userspace, since we know we're 
-        // not using that yet. It'll go at the 512GiB mark.
-        //
-        // Because we need mappings to be naturally aligned, we have to round the 
-        // phys address down to the next terapage (512GiB) boundary - which will
-        // almost certainly make it 0 on any system we're likely to be running on.
-        //
-        // We then map that in at the 512GiB mark (to avoid accidental nulls
-        // creeping in somewhere) and offset that by the phys address' offset
-        // from its next-lowest 512GiB boundary (as I say, likely zero) so
-        // we can get at the table.
-        //
-        // This is a bit painful, but saves some hassle, since we can't
-        // go the nicer way until this direct mapping is set up...
-        //
-
-        // Figure out the next-lowest 512GiB base and the offset of the phys address...
-        const uintptr_t pdpt_phys = vmm_table_entry_to_phys(pml4[pml4e]);
-        const uintptr_t pdpt_phys_base_addr = round_down_to_terapage(pdpt_phys);
-        const uintptr_t pdpt_phys_addr_offs = offset_in_terapage(pdpt_phys);
-
-        // This is the base of the virtual space we're about to map...
-        const uintptr_t pdpt_temp_vaddr = TERA_PAGE_SIZE;
-
-        // Create a pointer we'll use to get at the table once we've mapped it...
-        uint64_t * const temp_mapped_pdpt = (uint64_t*)(pdpt_temp_vaddr + pdpt_phys_addr_offs);
-
-        // Okay, good - do the mapping & flush TLB.
-        map_readwrite_and_flush_offs(pml4, 1, pdpt_phys_base_addr, pdpt_temp_vaddr, pdpt_phys_addr_offs);
-
-        // Right, now we have a PDPT mapped with a virtual pointer we can use it...
-        uint16_t pdpte = vmm_virt_to_pdpt_index(temp_map_addr);
-
-        vdebugf("  Will try to get entry %d from the pdpt\n", pdpte);
-        if (unlikely((temp_mapped_pdpt[pdpte] & PG_PRESENT) == 0)) {
-            vdebugf("    PD is not present, will create PD and PT\n");
-
-            // We need to create a PD and a PT, let's do that. We'll use the 
-            // next 512GiB of userspace as our temporary mapping. This probably 
-            // isn't necessary, since the chances are both these physical pages
-            // will be below 512GiB physical, but assuming that will be great
-            // until some crazy system comes along with >512GiB physical RAM...
-            //
-            // So let's not assume :D
-            //
-            // (And yeah, we _could_ set the PDPT entry before remapping these 
-            // to zero them, but there's a chance that could lead to TLB poisoning
-            // due to speculation, so we'll map and zero them first then hook it
-            // up...)
-            //
-            uintptr_t pd_phys = page_alloc(physical_region);
-            uintptr_t pt_phys = page_alloc(physical_region);
-
-            // #### Handle the page directory...
-            //
-            // This is essentially the same as we did above... figure out the next-lowest
-            // 512GiB base and the offset of the phys address, for the PD to begin with.
-            uintptr_t temp_table_phys = pd_phys;
-            uintptr_t temp_table_phys_base_addr = round_down_to_terapage(temp_table_phys);
-            uintptr_t temp_table_phys_addr_offs = offset_in_terapage(temp_table_phys);
-
-            // This is the base of the virtual space we're about to map...
-            uintptr_t temp_table_vaddr = TERA_PAGE_SIZE * 2;
-
-            // Create a pointer we'll use to get at the table once we've mapped it...
-            uint64_t* temp_mapped_table = (uint64_t*)(temp_table_vaddr + temp_table_phys_addr_offs);
-
-            // Do do the mapping & flush TLB.
-            map_readwrite_and_flush_offs(pml4, 4, temp_table_phys_base_addr, temp_table_vaddr, temp_table_phys_addr_offs);
-
-            // Zero the page, and map in the PT...
-            for (int i = 0; i < PAGE_TABLE_ENTRIES; i++) {
-                temp_mapped_table[i] = 0;
+                spinlock_unlock_irqrestore(&vmm_map_lock, lock_flags);
+                return vmm_table_entry_to_phys(pdpte);
             }
 
-            // Right, now we have a PD mapped with a virtual pointer we can use it...
-            uint16_t pde = vmm_virt_to_pd_index(temp_map_addr);
-            temp_mapped_table[pde] = vmm_phys_and_flags_to_table_entry(pt_phys, PG_PRESENT);
+            uint64_t *pd = vmm_phys_to_virt_ptr(vmm_table_entry_to_phys(pdpte));
+            uint16_t pd_index = vmm_virt_to_pd_index(virt_addr);
+            uint64_t pde = pd[pd_index];
 
-            // #### Handle the page table...
-            //
-            // This is the same thing again, but I'm not refactoring out the commonality
-            // since it's easier to follow this way - it's tricky enough without trying
-            // to be clever (so I'll let the compiler inject cleverness :D)
-            temp_table_phys = pt_phys;
-            temp_table_phys_base_addr = round_down_to_terapage(temp_table_phys);
-            temp_table_phys_addr_offs = offset_in_terapage(temp_table_phys);
+            if (pde & PG_PRESENT) {
+                if (is_leaf(pde)) {
+                    // unmapping a megapage
+                    pd[pd_index] = 0;
+                    vmm_invalidate_page(virt_addr);
 
-            // Do do the mapping & flush TLB.
-            map_readwrite_and_flush_offs(pml4, 2, temp_table_phys_base_addr, temp_table_vaddr, temp_table_phys_addr_offs);
-
-            // Just zero this one, that's all we need
-            for (int i = 0; i < PAGE_TABLE_ENTRIES; i++) {
-                temp_mapped_table[i] = 0;
-            }
-
-            // Finally, map the PD into the PDPT & invalidate
-            temp_mapped_pdpt[pdpte] = vmm_phys_and_flags_to_table_entry(pd_phys, PG_PRESENT);
-            cpu_invalidate_tlb_addr(temp_map_addr);
-
-            // Remove the temp PDPT mapping in the PML4
-            // 
-            // We're returning the one at pml4[2], so we don't clean that 
-            // up - that'll be handled in cleanup_cpu0_temp_page_tables later.
-            pml4[1] = 0;
-            cpu_invalidate_tlb_addr(pdpt_temp_vaddr + pdpt_phys_addr_offs);
-
-            // and return the PT vaddr
-            vdebugf("ensure_temp_page_tables(0x%016lx, 0x%016lx): success, return 0x%016lx\n", (uintptr_t)pml4, temp_map_addr, (uintptr_t)temp_mapped_table);
-            return temp_mapped_table;            
-        } else {
-            // So we maybe just need a PT, fine. 
-            //
-            // We'll use the next 512GiB of userspace as our temporary mapping
-            // for this one (i.e. mapping at the 1TiB mark). 
-            //
-            vdebugf("    PD is present, checking PT\n");
-
-            // First let's get hold of the PD
-
-            // Figure out the next-lowest 512GiB base and the offset of the phys address...
-            const uintptr_t pd_phys = vmm_table_entry_to_phys(temp_mapped_pdpt[pdpte]);
-            const uintptr_t pd_phys_base_addr = round_down_to_terapage(pd_phys);
-            const uintptr_t pd_phys_addr_offs = offset_in_terapage(pd_phys);
-
-            // This is the base of the virtual space we're about to map...
-            const uintptr_t pd_temp_vaddr = TERA_PAGE_SIZE * 2;
-
-            // Create a pointer we'll use to get at the table once we've mapped it...
-            uint64_t * const temp_mapped_pd = (uint64_t*)(pd_temp_vaddr + pd_phys_addr_offs);
-
-            // Okay, good - do the mapping & flush TLB.
-            map_readwrite_and_flush_offs(pml4, 2, pd_phys_base_addr, pd_temp_vaddr, pd_phys_addr_offs);
-
-            // Right, now we have a PDPT mapped with a virtual pointer we can use it...
-            uint16_t pde = vmm_virt_to_pd_index(temp_map_addr);
-
-            vdebugf("  Will try to get entry %d from the pd\n", pde);
-
-            if (likely((temp_mapped_pd[pde] & PG_PRESENT) == 0)) {
-                vdebugf("    PT is not present, will create one\n");
-
-                uintptr_t pt_phys = page_alloc(physical_region);
-
-                // #### Handle the page table...
-                //
-                // This is the same thing again, but I'm not refactoring out the commonality
-                // since it's easier to follow this way - it's tricky enough without trying
-                // to be clever (so I'll let the compiler inject cleverness :D)
-                const uintptr_t temp_table_phys = pt_phys;
-                const uintptr_t temp_table_phys_base_addr = round_down_to_terapage(temp_table_phys);
-                const uintptr_t temp_table_phys_addr_offs = offset_in_terapage(temp_table_phys);
-
-                // This is the base of the virtual space we're about to map...
-                uintptr_t temp_table_vaddr = TERA_PAGE_SIZE * 3;
-
-                // Create a pointer we'll use to get at the table once we've mapped it...
-                uint64_t* temp_mapped_table = (uint64_t*)(temp_table_vaddr + temp_table_phys_addr_offs);
-                
-                // Do do the mapping & flush TLB.
-                map_readwrite_and_flush_offs(pml4, 3, temp_table_phys_base_addr, temp_table_vaddr, temp_table_phys_addr_offs);
-    
-                // Just zero this one, that's all we need
-                for (int i = 0; i < PAGE_TABLE_ENTRIES; i++) {
-                    temp_mapped_table[i] = 0;
+                    spinlock_unlock_irqrestore(&vmm_map_lock, lock_flags);
+                    return vmm_table_entry_to_phys(pde);
                 }
-    
-                // Finally, map the PT into the PDPT & invalidate
-                temp_mapped_pd[pde] = vmm_phys_and_flags_to_table_entry(pt_phys, PG_PRESENT);
-                cpu_invalidate_tlb_addr(temp_map_addr);
 
-                // Remove the temp PDPT and PD mappings in the PML4
-                // 
-                // We're returning the one at pml4[3], so we don't clean that 
-                // up - that'll be handled in cleanup_cpu0_temp_page_tables later.
-                pml4[1] = 0;
-                cpu_invalidate_tlb_addr(pdpt_temp_vaddr + pdpt_phys_addr_offs);
-                pml4[2] = 0;
-                cpu_invalidate_tlb_addr(pd_temp_vaddr + pd_phys_addr_offs);
+                uint64_t *pt =
+                        vmm_phys_to_virt_ptr(vmm_table_entry_to_phys(pde));
+                uint16_t pt_index = vmm_virt_to_pt_index(virt_addr);
+                uint64_t pte = pt[pt_index];
 
-                // and return the PT vaddr
-                vdebugf("ensure_temp_page_tables(0x%016lx, 0x%016lx): success, return 0x%016lx\n", (uintptr_t)pml4, temp_map_addr, (uintptr_t)temp_mapped_table);
-                return temp_mapped_table;
-            } else {
-                // I see, we don't need anything. This is a potential worry (we might've initialised
-                // things in the wrong order) but fine, we'll map a vaddr for the PT and just return
-                // a pointer to it.
+                if (pte & PG_PRESENT) {
+                    if (is_leaf(pte)) {
+                        // unmapping a megapage
+                        pt[pt_index] = 0;
 
-                vdebugf("    PT is present, will just map it temporarily\n");
+                        vmm_invalidate_page(virt_addr);
 
-                // Figure out the next-lowest 512GiB base and the offset of the phys address...
-                const uintptr_t pt_phys = vmm_table_entry_to_phys(temp_mapped_pd[pde]);
-                const uintptr_t pt_phys_base_addr = round_down_to_terapage(pt_phys);
-                const uintptr_t pt_phys_addr_offs = offset_in_terapage(pt_phys);
-
-                // This is the base of the virtual space we're about to map...
-                const uintptr_t pt_temp_vaddr = TERA_PAGE_SIZE * 2;
-
-                // Create a pointer we'll use to get at the table once we've mapped it...
-                uint64_t * const temp_mapped_pt = (uint64_t*)(pt_temp_vaddr + pt_phys_addr_offs);
-
-                // Okay, good - do the mapping & flush TLB.
-                // We'll just replace the temp mapped PD at this point, we're done with it...
-                map_readwrite_and_flush_offs(pml4, 2, pt_phys_base_addr, pt_temp_vaddr, pt_phys_addr_offs);
-
-                // Remove the temp PDPT mapping in the PML4
-                // 
-                // We're returning the one at pml4[2], so we don't clean that 
-                // up - that'll be handled in cleanup_cpu0_temp_page_tables later.
-                pml4[1] = 0;
-                cpu_invalidate_tlb_addr(pdpt_temp_vaddr + pdpt_phys_addr_offs);
-
-                // and return the PT vaddr
-                vdebugf("ensure_temp_page_tables(0x%016lx, 0x%016lx): success, return 0x%016lx\n", (uintptr_t)pml4, temp_map_addr, (uintptr_t)temp_mapped_pt);
-                return temp_mapped_pt;
+                        spinlock_unlock_irqrestore(&vmm_map_lock, lock_flags);
+                        return vmm_table_entry_to_phys(pte);
+                    }
+                }
             }
         }
     }
 
-    vdebugf("ensure_temp_page_tables(0x%016lx, 0x%016lx): failed, return NULL\n", (uintptr_t)pml4, temp_map_addr);
+    spinlock_unlock_irqrestore(&vmm_map_lock, lock_flags);
     return 0;
 }
 
-// clang-format on
-
-void cleanup_temp_page_tables(uint64_t *const pml4) {
-    // This _could_ go about walking the tables and cleaning up exactly
-    // what we ended up mapping in ensure_temp_page_tables and during
-    // the direct map build itself, but since we'll do this exactly
-    // once at the end of the whole direct-mapping process, let's just
-    // take the easy way out and remove all the userspace mappings we
-    // might've fiddled with and dump the whole TLB so we start afresh.
-
-    pml4[1] = 0;
-    pml4[2] = 0;
-    pml4[3] = 0;
-
-    // **boom**
-    cpu_invalidate_tlb_all();
+inline uintptr_t vmm_unmap_page(uintptr_t virt_addr) {
+    return vmm_unmap_page_in(
+            vmm_phys_to_virt_ptr(cpu_satp_to_root_table_phys(cpu_read_satp())),
+            virt_addr);
 }
-
-void vmm_init_direct_mapping(uint64_t *pml4, Limine_MemMap *memmap) {
-    vdebugf("vmm_init_direct_mapping: init with %ld entries at pml4 0x%016lx\n",
-            memmap->entry_count, (uintptr_t)pml4);
-
-    uint64_t *temp_pt =
-            ensure_temp_page_tables(pml4, vmm_per_cpu_temp_page_addr(0));
-
-    vdebugf("ensure tables returns 0x%016lx\n", (uintptr_t)temp_pt);
-
-    for (int i = 0; i < memmap->entry_count; i++) {
-        switch (memmap->entries[i]->type) {
-        case LIMINE_MEMMAP_USABLE:
-        case LIMINE_MEMMAP_BOOTLOADER_RECLAIMABLE:
-        case LIMINE_MEMMAP_ACPI_RECLAIMABLE:
-            vmm_init_map_region(pml4, temp_pt, memmap->entries[i], true);
-            break;
-        case LIMINE_MEMMAP_ACPI_NVS:
-        case LIMINE_MEMMAP_EXECUTABLE_AND_MODULES:
-            vmm_init_map_region(pml4, temp_pt, memmap->entries[i], false);
-            break;
-        default:
-            vdebugf("vmm_init_direct_mapping: ignored region with type %ld\n",
-                    memmap->entries[i]->type);
-        }
-    }
-
-    cleanup_temp_page_tables(pml4);
-}
-
-// TODO The following is mostly bullshit, to be removed.
-//
-
-// // Initialize the direct mapping for physical memory
-// void vmm_init_direct_mapping(void) {
-//     // Allocate the top-level page table
-//     current_hierarchy.l2 = vmm_alloc_page_table();
-//     if (!current_hierarchy.l2) {
-//         // Handle error
-//         return;
-//     }
-
-// #ifdef RISCV_SV48
-//     // For Sv48, we need to allocate L3 as well
-//     current_hierarchy.l3 = vmm_alloc_page_table();
-//     if (!current_hierarchy.l3) {
-//         // Handle error
-//         return;
-//     }
-// #endif
-
-//     // Set up the direct mapping for physical memory
-//     // We'll use huge pages (2MB) for efficiency
-//     uintptr_t phys_addr = 0;
-//     uintptr_t virt_addr = DIRECT_MAP_BASE;
-
-//     // Map all physical memory using huge pages
-//     // This is a simplified version - in a real implementation,
-//     // you would get the actual physical memory size from the bootloader
-//     // or memory manager
-//     while (phys_addr < 0x100000000) { // 4GB for this example
-//         vmm_map_page(virt_addr, phys_addr,
-//                      PG_PRESENT | PG_READ | PG_WRITE | PG_GLOBAL);
-//         phys_addr += MEGA_PAGE_SIZE;
-//         virt_addr += MEGA_PAGE_SIZE;
-//     }
-
-//     // Set the page table base address in the SATP CSR
-//     uintptr_t pt_base = virt_to_phys((uintptr_t)current_hierarchy.l2);
-// #ifdef RISCV_SV48
-//     cpu_set_satp(pt_base, SATP_MODE_SV48);
-// #else
-//     cpu_set_satp(pt_base, SATP_MODE_SV39);
-// #endif
-// }
-
-// // Get the page table entry for a given virtual address
-// uint64_t *vmm_get_pte(uintptr_t virt_addr) {
-//     // Extract the page table indices from the virtual address
-//     uint16_t l0_index = (virt_addr >> 12) & 0x1FF;
-//     uint16_t l1_index = (virt_addr >> 21) & 0x1FF;
-//     uint16_t l2_index = (virt_addr >> 30) & 0x1FF;
-// #ifdef RISCV_SV48
-//     uint16_t l3_index = (virt_addr >> 39) & 0x1FF;
-// #endif
-
-//     // Get the physical address of the page table
-//     uintptr_t pt_phys_addr;
-
-// #ifdef RISCV_SV48
-//     // For Sv48, we need to traverse L3, L2, L1, L0
-//     if (!current_hierarchy.l3)
-//         return NULL;
-
-//     // Get L3 entry
-//     uint64_t l3_entry = current_hierarchy.l3->entries[l3_index];
-//     if (!(l3_entry & PG_PRESENT))
-//         return NULL;
-
-//     // Get L2
-//     PageTable *l2 = (PageTable *)phys_to_virt(l3_entry & ~0xFFF);
-//     if (!l2)
-//         return NULL;
-
-//     // Get L2 entry
-//     uint64_t l2_entry = l2->entries[l2_index];
-//     if (!(l2_entry & PG_PRESENT))
-//         return NULL;
-
-//     // Get L1
-//     PageTable *l1 = (PageTable *)phys_to_virt(l2_entry & ~0xFFF);
-//     if (!l1)
-//         return NULL;
-
-//     // Get L1 entry
-//     uint64_t l1_entry = l1->entries[l1_index];
-//     if (!(l1_entry & PG_PRESENT))
-//         return NULL;
-
-//     // Get L0
-//     PageTable *l0 = (PageTable *)phys_to_virt(l1_entry & ~0xFFF);
-//     if (!l0)
-//         return NULL;
-
-//     // Return L0 entry
-//     return &l0->entries[l0_index];
-// #else
-//     // For Sv39, we need to traverse L2, L1, L0
-//     if (!current_hierarchy.l2)
-//         return NULL;
-
-//     // Get L2 entry
-//     uint64_t l2_entry = current_hierarchy.l2->entries[l2_index];
-//     if (!(l2_entry & PG_PRESENT))
-//         return NULL;
-
-//     // Get L1
-//     PageTable *l1 = (PageTable *)phys_to_virt(l2_entry & ~0xFFF);
-//     if (!l1)
-//         return NULL;
-
-//     // Get L1 entry
-//     uint64_t l1_entry = l1->entries[l1_index];
-//     if (!(l1_entry & PG_PRESENT))
-//         return NULL;
-
-//     // Get L0
-//     PageTable *l0 = (PageTable *)phys_to_virt(l1_entry & ~0xFFF);
-//     if (!l0)
-//         return NULL;
-
-//     // Return L0 entry
-//     return &l0->entries[l0_index];
-// #endif
-// }
-
-// // Get the page table for a given virtual address
-// PageTable *vmm_get_pt(uintptr_t virt_addr) {
-//     // Extract the page table indices from the virtual address
-//     uint16_t l1_index = (virt_addr >> 21) & 0x1FF;
-//     uint16_t l2_index = (virt_addr >> 30) & 0x1FF;
-// #ifdef RISCV_SV48
-//     uint16_t l3_index = (virt_addr >> 39) & 0x1FF;
-// #endif
-
-// #ifdef RISCV_SV48
-//     // For Sv48, we need to traverse L3, L2, L1
-//     if (!current_hierarchy.l3)
-//         return NULL;
-
-//     // Get L3 entry
-//     uint64_t l3_entry = current_hierarchy.l3->entries[l3_index];
-//     if (!(l3_entry & PG_PRESENT))
-//         return NULL;
-
-//     // Get L2
-//     PageTable *l2 = (PageTable *)phys_to_virt(l3_entry & ~0xFFF);
-//     if (!l2)
-//         return NULL;
-
-//     // Get L2 entry
-//     uint64_t l2_entry = l2->entries[l2_index];
-//     if (!(l2_entry & PG_PRESENT))
-//         return NULL;
-
-//     // Get L1
-//     PageTable *l1 = (PageTable *)phys_to_virt(l2_entry & ~0xFFF);
-//     if (!l1)
-//         return NULL;
-
-//     // Get L1 entry
-//     uint64_t l1_entry = l1->entries[l1_index];
-//     if (!(l1_entry & PG_PRESENT))
-//         return NULL;
-
-//     // Get L0
-//     return (PageTable *)phys_to_virt(l1_entry & ~0xFFF);
-// #else
-//     // For Sv39, we need to traverse L2, L1
-//     if (!current_hierarchy.l2)
-//         return NULL;
-
-//     // Get L2 entry
-//     uint64_t l2_entry = current_hierarchy.l2->entries[l2_index];
-//     if (!(l2_entry & PG_PRESENT))
-//         return NULL;
-
-//     // Get L1
-//     PageTable *l1 = (PageTable *)phys_to_virt(l2_entry & ~0xFFF);
-//     if (!l1)
-//         return NULL;
-
-//     // Get L1 entry
-//     uint64_t l1_entry = l1->entries[l1_index];
-//     if (!(l1_entry & PG_PRESENT))
-//         return NULL;
-
-//     // Get L0
-//     return (PageTable *)phys_to_virt(l1_entry & ~0xFFF);
-// #endif
-// }
-
-// // Get the page directory for a given virtual address
-// PageTable *vmm_get_pd(uintptr_t virt_addr) {
-//     // Extract the page table indices from the virtual address
-//     uint16_t l2_index = (virt_addr >> 30) & 0x1FF;
-// #ifdef RISCV_SV48
-//     uint16_t l3_index = (virt_addr >> 39) & 0x1FF;
-// #endif
-
-// #ifdef RISCV_SV48
-//     // For Sv48, we need to traverse L3, L2
-//     if (!current_hierarchy.l3)
-//         return NULL;
-
-//     // Get L3 entry
-//     uint64_t l3_entry = current_hierarchy.l3->entries[l3_index];
-//     if (!(l3_entry & PG_PRESENT))
-//         return NULL;
-
-//     // Get L2
-//     PageTable *l2 = (PageTable *)phys_to_virt(l3_entry & ~0xFFF);
-//     if (!l2)
-//         return NULL;
-
-//     // Get L2 entry
-//     uint64_t l2_entry = l2->entries[l2_index];
-//     if (!(l2_entry & PG_PRESENT))
-//         return NULL;
-
-//     // Get L1
-//     return (PageTable *)phys_to_virt(l2_entry & ~0xFFF);
-// #else
-//     // For Sv39, we need to traverse L2
-//     if (!current_hierarchy.l2)
-//         return NULL;
-
-//     // Get L2 entry
-//     uint64_t l2_entry = current_hierarchy.l2->entries[l2_index];
-//     if (!(l2_entry & PG_PRESENT))
-//         return NULL;
-
-//     // Get L1
-//     return (PageTable *)phys_to_virt(l2_entry & ~0xFFF);
-// #endif
-// }
-
-// // Get the page directory pointer table for a given virtual address
-// PageTable *vmm_get_pdpt(uintptr_t virt_addr) {
-// #ifdef RISCV_SV48
-//     // Extract the page table indices from the virtual address
-//     uint16_t l3_index = (virt_addr >> 39) & 0x1FF;
-
-//     // For Sv48, we need to traverse L3
-//     if (!current_hierarchy.l3)
-//         return NULL;
-
-//     // Get L3 entry
-//     uint64_t l3_entry = current_hierarchy.l3->entries[l3_index];
-//     if (!(l3_entry & PG_PRESENT))
-//         return NULL;
-
-//     // Get L2
-//     return (PageTable *)phys_to_virt(l3_entry & ~0xFFF);
-// #else
-//     // For Sv39, L2 is the top level
-//     return current_hierarchy.l2;
-// #endif
-// }
-
-// #ifdef RISCV_SV48
-// // Get the page map level 4 for a given virtual address
-// PageTable *vmm_get_pml4(uintptr_t virt_addr) {
-//     // For Sv48, L3 is the top level
-//     return current_hierarchy.l3;
-// }
-// #endif
-
-// // Get the physical address for a given virtual address
-// uintptr_t vmm_virt_to_phys(uintptr_t virt_addr) {
-//     // Check if the address is in the direct mapping region
-//     if (is_direct_mapped(virt_addr)) {
-//         return virt_to_phys(virt_addr);
-//     }
-
-//     // Get the page table entry
-//     uint64_t *pte = vmm_get_pte(virt_addr);
-//     if (!pte || !(*pte & PG_PRESENT)) {
-//         return 0;
-//     }
-
-//     // Extract the physical address from the page table entry
-//     uintptr_t phys_addr = *pte & ~0xFFF;
-
-//     // Add the page offset
-//     phys_addr |= virt_addr & 0xFFF;
-
-//     return phys_addr;
-// }
-
-// // Map a physical page to a virtual address
-// bool vmm_map_page(uintptr_t virt_addr, uintptr_t phys_addr, uint64_t flags) {
-//     // Extract the page table indices from the virtual address
-//     uint16_t l0_index = (virt_addr >> 12) & 0x1FF;
-//     uint16_t l1_index = (virt_addr >> 21) & 0x1FF;
-//     uint16_t l2_index = (virt_addr >> 30) & 0x1FF;
-// #ifdef RISCV_SV48
-//     uint16_t l3_index = (virt_addr >> 39) & 0x1FF;
-// #endif
-
-//     // Ensure the page tables exist
-// #ifdef RISCV_SV48
-//     // For Sv48, we need to ensure L3, L2, L1, L0 exist
-//     if (!current_hierarchy.l3) {
-//         current_hierarchy.l3 = vmm_alloc_page_table();
-//         if (!current_hierarchy.l3)
-//             return false;
-//     }
-
-//     // Get L3 entry
-//     uint64_t l3_entry = current_hierarchy.l3->entries[l3_index];
-//     if (!(l3_entry & PG_PRESENT)) {
-//         // Allocate L2
-//         PageTable *l2 = vmm_alloc_page_table();
-//         if (!l2)
-//             return false;
-
-//         // Set L3 entry
-//         current_hierarchy.l3->entries[l3_index] =
-//                 ((uintptr_t)virt_to_phys((uintptr_t)l2) & ~0xFFF) | PG_PRESENT;
-//     }
-
-//     // Get L2
-//     PageTable *l2 = (PageTable *)phys_to_virt(
-//             current_hierarchy.l3->entries[l3_index] & ~0xFFF);
-
-//     // Get L2 entry
-//     uint64_t l2_entry = l2->entries[l2_index];
-//     if (!(l2_entry & PG_PRESENT)) {
-//         // Allocate L1
-//         PageTable *l1 = vmm_alloc_page_table();
-//         if (!l1)
-//             return false;
-
-//         // Set L2 entry
-//         l2->entries[l2_index] =
-//                 ((uintptr_t)virt_to_phys((uintptr_t)l1) & ~0xFFF) | PG_PRESENT;
-//     }
-
-//     // Get L1
-//     PageTable *l1 = (PageTable *)phys_to_virt(l2->entries[l2_index] & ~0xFFF);
-
-//     // Get L1 entry
-//     uint64_t l1_entry = l1->entries[l1_index];
-//     if (!(l1_entry & PG_PRESENT)) {
-//         // Allocate L0
-//         PageTable *l0 = vmm_alloc_page_table();
-//         if (!l0)
-//             return false;
-
-//         // Set L1 entry
-//         l1->entries[l1_index] =
-//                 ((uintptr_t)virt_to_phys((uintptr_t)l0) & ~0xFFF) | PG_PRESENT;
-//     }
-
-//     // Get L0
-//     PageTable *l0 = (PageTable *)phys_to_virt(l1->entries[l1_index] & ~0xFFF);
-
-//     // Set L0 entry
-//     l0->entries[l0_index] = (phys_addr & ~0xFFF) | flags;
-// #else
-//     // For Sv39, we need to ensure L2, L1, L0 exist
-//     if (!current_hierarchy.l2) {
-//         current_hierarchy.l2 = vmm_alloc_page_table();
-//         if (!current_hierarchy.l2)
-//             return false;
-//     }
-
-//     // Get L2 entry
-//     uint64_t l2_entry = current_hierarchy.l2->entries[l2_index];
-//     if (!(l2_entry & PG_PRESENT)) {
-//         // Allocate L1
-//         PageTable *l1 = vmm_alloc_page_table();
-//         if (!l1)
-//             return false;
-
-//         // Set L2 entry
-//         current_hierarchy.l2->entries[l2_index] =
-//                 ((uintptr_t)virt_to_phys((uintptr_t)l1) & ~0xFFF) | PG_PRESENT;
-//     }
-
-//     // Get L1
-//     PageTable *l1 = (PageTable *)phys_to_virt(
-//             current_hierarchy.l2->entries[l2_index] & ~0xFFF);
-
-//     // Get L1 entry
-//     uint64_t l1_entry = l1->entries[l1_index];
-//     if (!(l1_entry & PG_PRESENT)) {
-//         // Allocate L0
-//         PageTable *l0 = vmm_alloc_page_table();
-//         if (!l0)
-//             return false;
-
-//         // Set L1 entry
-//         l1->entries[l1_index] =
-//                 ((uintptr_t)virt_to_phys((uintptr_t)l0) & ~0xFFF) | PG_PRESENT;
-//     }
-
-//     // Get L0
-//     PageTable *l0 = (PageTable *)phys_to_virt(l1->entries[l1_index] & ~0xFFF);
-
-//     // Set L0 entry
-//     l0->entries[l0_index] = (phys_addr & ~0xFFF) | flags;
-// #endif
-
-//     // Invalidate the TLB for this address
-//     cpu_invalidate_tlb_addr(virt_addr);
-
-//     return true;
-// }
-
-// // Unmap a virtual address
-// bool vmm_unmap_page(uintptr_t virt_addr) {
-//     // Get the page table entry
-//     uint64_t *pte = vmm_get_pte(virt_addr);
-//     if (!pte || !(*pte & PG_PRESENT)) {
-//         return false;
-//     }
-
-//     // Clear the page table entry
-//     *pte = 0;
-
-//     // Invalidate the TLB for this address
-//     cpu_invalidate_tlb_addr(virt_addr);
-
-//     return true;
-// }
-
-// // Allocate a new page table
-// PageTable *vmm_alloc_page_table(void) {
-//     // Allocate a physical page for the page table
-//     uintptr_t phys_addr = page_alloc(physical_region);
-//     if (!phys_addr) {
-//         return NULL;
-//     }
-
-//     // Convert to virtual address
-//     PageTable *pt = (PageTable *)phys_to_virt(phys_addr);
-
-//     // Clear the page table
-//     memset(pt, 0, sizeof(PageTable));
-
-//     return pt;
-// }
-
-// // Free a page table
-// void vmm_free_page_table(PageTable *pt) {
-//     if (!pt) {
-//         return;
-//     }
-
-//     // Convert to physical address
-//     uintptr_t phys_addr = virt_to_phys((uintptr_t)pt);
-
-//     // Free the physical page
-//     page_free(physical_region, phys_addr);
-// }
