+++ kernel/entrypoint.c
@@ -145,7 +145,7 @@ noreturn void ap_kernel_entrypoint(uint64_t ap_num) {
 
 // Common entrypoint once bootloader-specific stuff is handled
 noreturn void bsp_kernel_entrypoint(uintptr_t rsdp_phys) {
-    if (!fba_init((uint64_t *)vmm_recursive_find_pml4(), KERNEL_FBA_BEGIN,
+    if (!fba_init((uint64_t *)vmm_find_pml4(), KERNEL_FBA_BEGIN,
                   KERNEL_FBA_SIZE_BLOCKS)) {
         panic("FBA init failed");
     }
