+++ kernel/tests/mock_vmm.c
@@ -57,8 +57,7 @@ bool vmm_map_page_in(void *pml4, uintptr_t virt_addr, uint64_t page,
 }
 
 bool vmm_map_page(uintptr_t virt_addr, uint64_t page, uint16_t flags) {
-    return vmm_map_page_in((uint64_t *)vmm_recursive_find_pml4(), virt_addr,
-                           page, flags);
+    return vmm_map_page_in((uint64_t *)vmm_find_pml4(), virt_addr, page, flags);
 }
 
 uintptr_t vmm_unmap_page_in(uint64_t *pml4, uintptr_t virt_addr) {
@@ -76,5 +75,5 @@ bool vmm_map_page_containing(uintptr_t virt_addr, uint64_t phys_addr,
 }
 
 uintptr_t vmm_unmap_page(uintptr_t virt_addr) {
-    return vmm_unmap_page_in((uint64_t *)vmm_recursive_find_pml4(), virt_addr);
+    return vmm_unmap_page_in((uint64_t *)vmm_find_pml4(), virt_addr);
 }
